tools/requirements.txt
python-dotenv
boto3
ibm-watson
azure-cognitiveservices-speech
google-cloud-texttospeech